Simplify and rename Go SDK LLM callbacks to CopilotRequestHandler
Mirror the .NET/Node simplification + terminology rename in the Go SDK:
consolidate the provider/handler two-layer design into a single
copilot_request_handler.go, drop the accepted:false ack plumbing and the
staged backstop, emit the WebSocket 101 upgrade head eagerly (a lazy bridge
deadlocks the runtime connect), and rename the public Llm* types to Copilot*
(types carry the prefix; fields/methods stay succinct). The client option
becomes ClientOptions.RequestHandler *CopilotRequestHandler. Generated wire
types are untouched.

Consolidate the e2e suite to three files (copilot_request_handler covering
HTTP + WebSocket + streaming, copilot_request_session_id, and
copilot_request_cancel_error with the error and runtime-cancel cases) plus a
shared helpers file, replacing the nine llm_inference_* test files.
